02-device_prediction/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "device-prediction"
version = "0.1.0"
description = "Device change prediction using survival analysis"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-bigquery>=3.0",
    "db-dtypes>=1.0",
    "google-genai>=1.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "scikit-survival>=0.23",
    "joblib>=1.3",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "tenacity>=8.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov"]
xgboost = ["xgboost>=2.0"]

[project.scripts]
device-prediction = "device_prediction.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
